+++ src/events.h
@@ -1,6 +1,6 @@
 #include "common.h"
 
-void do_event(s32 event);
+void do_event(u8 event);
 
 extern void do_screenshot(void);
