+++ src/init.h
@@ -9,6 +9,6 @@ void init_buttons(TTF_Font *font);
 void init_keybinds(void);
 
 extern GPU_Image *load_texture(char *filename);
-extern GPU_Image* textureFromFont(TTF_Font *font, char *text, s32 style, SDL_Color colour);
+extern GPU_Image* textureFromFont(TTF_Font *font, char *text, u8 style, SDL_Color colour);
 
 extern App app;
