+++ src/input.h
@@ -5,6 +5,6 @@ void do_key_down(SDL_KeyboardEvent *event);
 void do_input(void);
 bool check_keypress(u8 action, u8 keybind);
 
-extern void do_event(s32 input);
+extern void do_event(u8 input);
 
 extern App app;
